Fix inconsistent matrix algebra between OpenGL and CGAffineTransform
It had previously escaped me that definitions of the operations on the
touchHLE internal `Matrix` type and OpenGL transformation matrices
are transposed relative to their CGAffineTransform counterparts.
This led to incorrect partial fixes in
8227159cf2d5ab4542e53c121f0b06f600a29a6b and
183d8cf43e8e8cdd472f66aff540004f4581ab6e which made the `Matrix` type's
operations internally inconsistent, and introduced hacky workarounds.

This commit comprehensively fixes this and removes the workarounds.
`Matrix` once again consistently uses the OpenGL style, and the mismatch
with CGAffineTransform is resolved by transposing when converting
between the two types.

Change-Id: Ie58c223c1d3ed4145f42fe3e49e85336eeb489bb